move errorMessage from fooks to utils
